# server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: test_ex_ts
  description: My cool TypeScript app
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
...
parameters:
  schoolName:
    name: schoolName
    in: path
    description: Name of a school
    required: true
    type: string 
...
paths:
...
  /schools/{schoolName}:
    get:
      tags:
        - Schools
      description: Get a school by name
      parameters:
        - $ref: "#/parameters/schoolName"
      responses:
        200:
          description: Get a school by name
...